修改了数据库的结构，适当调整接单的接口逻辑

+++ backend/src/main/java/com/autocrowd/backend/dto/order/CurrentOrderResponse.java
@@ -26,6 +26,6 @@ public class CurrentOrderResponse {
     @Data
     public static class VehicleInfoDTO {
         private String licensePlate;
-        private BigDecimal fuelLevel;
+        private Integer fuelLevel;
     }
 }
